ppt2yt/requirements.txt
# PPTX to YouTube Video Generation System - Dependencies

# PPTX Processing
python-pptx==0.6.21

# AI & Machine Learning
openai==1.97.1
requests==2.31.0

# Image Processing
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78

# Video Processing
ffmpeg-python==0.2.0

# YouTube API
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Configuration & Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
psutil==5.9.5

# Logging & Error Handling
loguru==0.7.3

# PowerPoint Direct Capture
pywin32==306
pdf2image==1.16.3

# Development (Optional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0

# Additional Dependencies
colorama==0.4.6
win32-setctime==1.2.0 